many updates, see desc.
* added distinct(separator) transform
* re-named Process.System to Process.ReadOnly and set accordingly
* updated Quartz.net
* updated Npgsql driver
* updated NLog
* updated Elasticsearch client
* Updated System.Net.Http
* Removed remaining WithDefaults obsolete method
* updated OrchardCMS module

+++ Pipeline.Provider.Ado/AdoEntityWriter.cs
@@ -46,7 +46,7 @@ namespace Transformalize.Provider.Ado {
 
                 var inserts = new List<IRow>(_output.Entity.InsertSize);
                 var updates = new List<IRow>(_output.Entity.InsertSize);
-                var batchCount = 0;
+                var batchCount = (uint)0;
 
                 if (_output.Entity.IsFirstRun) {
                     foreach (var row in part) {
